requirements.txt
flask
moviepy
flask-cors
openai
gtts
mutagen
pillow